requirements.txt
streamlit
langchain
langchain-chroma
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
python-dotenv
pypdf
ollama
sentence-transformers